Move and rename `isaric_raw`
