+++ ehrql/tables/beta/raw/tpp.py
@@ -7,4 +7,257 @@ The data provided by this schema are minimally transformed. They are very close 
 data provided by the underlying database tables. They are provided for data development
 and data curation purposes.
 """
-__all__ = []
+import datetime
+
+from ehrql.tables import Constraint, EventFrame, Series, table
+
+
+__all__ = [
+    "isaric",
+]
+
+
+class isaric_raw(EventFrame):
+    """
+    A subset of the ISARIC data.
+
+    These columns are deliberately all taken as strings while in a preliminary phase.
+    They will later change to more appropriate data types.
+
+    Descriptions taken from: [CCP_REDCap_ISARIC_data_dictionary_codebook.pdf][isaric_ddc_pdf]
+
+    [isaric_ddc_pdf]: https://github.com/isaric4c/wiki/blob/d6b87d59a277cf2f6deedeb5e8c1a970dbb970a3/ISARIC/CCP_REDCap_ISARIC_data_dictionary_codebook.pdf
+    """
+
+    # Demographics
+    age = Series(
+        str,
+        description="Age",
+    )
+    age_factor = Series(
+        str,
+        description="TODO",
+    )
+    calc_age = Series(
+        str,
+        description="Calculated age (comparing date of birth with date of enrolment). May be inaccurate if a date of February 29 is used.",
+    )
+    sex = Series(
+        str,
+        description="Sex at birth.",
+    )
+    ethnic___1 = Series(
+        str,
+        description="Ethnic group: Arab.",
+    )
+    ethnic___2 = Series(
+        str,
+        description="Ethnic group: Black.",
+    )
+    ethnic___3 = Series(
+        str,
+        description="Ethnic group: East Asian.",
+    )
+    ethnic___4 = Series(
+        str,
+        description="Ethnic group: South Asian.",
+    )
+    ethnic___5 = Series(
+        str,
+        description="Ethnic group: West Asian.",
+    )
+    ethnic___6 = Series(
+        str,
+        description="Ethnic group: Latin American.",
+    )
+    ethnic___7 = Series(
+        str,
+        description="Ethnic group: White.",
+    )
+    ethnic___8 = Series(
+        str,
+        description="Ethnic group: Aboriginal/First Nations.",
+    )
+    ethnic___9 = Series(
+        str,
+        description="Ethnic group: Other.",
+    )
+    ethnic___10 = Series(
+        str,
+        description="Ethnic group: N/A.",
+    )
+
+    # Vaccination
+    covid19_vaccine = Series(
+        str,
+        description="Has the patient received a Covid-19 vaccine (open label licenced product)?",
+    )
+    covid19_vaccined = Series(
+        datetime.date,
+        description="Date first vaccine given (Covid-19) if known.",
+    )
+    covid19_vaccine2d = Series(
+        datetime.date,
+        description="Date second vaccine given (Covid-19) if known.",
+    )
+    covid19_vaccined_nk = Series(
+        str,
+        description="First vaccine given (Covid-19) but date not known.",
+    )
+
+    # Clinical
+    corona_ieorres = Series(
+        str,
+        description="Suspected or proven infection with pathogen of public health interest.",
+    )
+    # Note the coriona spelling here (compared with corona_ieorres).
+    # The difference exists both in the TPP ISARIC_New database table,
+    # and the ISARIC data dictionary.
+    coriona_ieorres2 = Series(
+        str,
+        description="Proven or high likelihood of infection with pathogen of public health interest.",
+    )
+    coriona_ieorres3 = Series(
+        str,
+        description="Proven infection with pathogen of public health interest.",
+    )
+    inflammatory_mss = Series(
+        str,
+        description="Adult or child who meets case definition for inflammatory multi-system syndrome (MIS-C/MIS-A).",
+    )
+    cestdat = Series(
+        datetime.date,
+        description="Onset date of first/earliest symptom.",
+    )
+
+    # Clinical characteristics
+    chrincard = Series(
+        str,
+        description="Chronic cardiac disease, including congenital heart disease (not hypertension).",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    hypertension_mhyn = Series(
+        str,
+        description="Hypertension (physician diagnosed).",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    chronicpul_mhyn = Series(
+        str,
+        description="Chronic pulmonary disease (not asthma).",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    asthma_mhyn = Series(
+        str,
+        description="Asthma (physician diagnosed).",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    renal_mhyn = Series(
+        str,
+        description="Chronic kidney disease.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    mildliver = Series(
+        str,
+        description="Mild liver disease.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    modliv = Series(
+        str,
+        description="Moderate or severe liver disease",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    chronicneu_mhyn = Series(
+        str,
+        description="Chronic neurological disorder.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    malignantneo_mhyn = Series(
+        str,
+        description="Malignant neoplasm.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    chronichaemo_mhyn = Series(
+        str,
+        description="Chronic haematologic disease.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    aidshiv_mhyn = Series(
+        str,
+        description="AIDS/HIV.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    obesity_mhyn = Series(
+        str,
+        description="Obesity (as defined by clinical staff).",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    diabetes_type_mhyn = Series(
+        str,
+        description="Diabetes and type.",
+        constraints=[Constraint.Categorical(["NO", "1", "2", "N/K"])],
+    )
+    diabetescom_mhyn = Series(
+        str,
+        description="Diabetes with complications.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    diabetes_mhyn = Series(
+        str,
+        description="Diabetes without complications.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    rheumatologic_mhyn = Series(
+        str,
+        description="Rheumatologic disorder.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    dementia_mhyn = Series(
+        str,
+        description="Dementia.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    malnutrition_mhyn = Series(
+        str,
+        description="Malnutrition.",
+        constraints=[Constraint.Categorical(["YES", "NO", "Unknown"])],
+    )
+    smoking_mhyn = Series(
+        str,
+        description="Smoking.",
+        constraints=[
+            Constraint.Categorical(["Yes", "Never Smoked", "Former Smoker", "N/K"])
+        ],
+    )
+
+    # Admission
+    hostdat = Series(
+        datetime.date,
+        description="Admission date at this facility.",
+    )
+    hooccur = Series(
+        str,
+        description="Transfer from other facility?",
+    )
+    hostdat_transfer = Series(
+        datetime.date,
+        description="Admission date at previous facility.",
+    )
+    hostdat_transfernk = Series(
+        str,
+        description="Admission date at previous facility not known.",
+    )
+    readm_cov19 = Series(
+        str,
+        description="Is the patient being readmitted with Covid-19?",
+    )
+    dsstdat = Series(
+        datetime.date,
+        description="Date of enrolment.",
+    )
+    dsstdtc = Series(
+        datetime.date,
+        description="Outcome date.",
+    )
+
+
+isaric = table(isaric_raw)
